backend/portfolio/requirements.txt
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.3.1
django-tinymce==4.0.0
djangorestframework==3.15.1
pillow==10.2.0
sqlparse==0.4.4
